pyproject.toml
# !/usr/bin/env toml
#
# === FILE META OPENING ===
# file: ./trend-analyzer/pyproject.toml
# role: configuration
# desc: minimal uv project configuration for trend analyzer skeleton
# === FILE META CLOSING ===

[project]
name = "trend-analyzer"
version = "0.1.0"
description = "AI-powered health insurance claims trend analysis (skeleton)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core data processing
    "pandas>=2.3.2",
    "numpy>=2.3.3",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.1.1",
    
    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    
    # Visualization
    "matplotlib>=3.10.6",
    "plotly>=6.3.0",
    
    # Output formats
    "python-docx>=1.2.0",
    
    # AI (optional - can be removed if not using)
    "openai-agents>=0.3.0",
    
    # Type safety
    "pydantic>=2.11.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trend_analyzer"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
